Update tfquaternion.py
Made some minor changes: When the quaternion class needs to create a new quaternion (say in inverse) the new quaternion will take the default dtype, instead of the parent's dtype, which will cause a type error if the parent does not use the default dtype.  I also added a utility function which I find to be the most convenient way of defining a quaternion when rotating a collection of vectors in the same way (i.e. rotating a mesh in CAD without distorting it).